+++ src/components/features/disposals/reject-disposal-dialog.tsx
@@ -3,7 +3,7 @@
 import { useState, useTransition } from 'react';
 import { AlertCircle } from 'lucide-react'; 
 
-import { rejectDisposalRequest } from '@/actions/disposals';
+import { rejectDisposalRequest } from '@/actions/disposals/reject';
 import { Button } from '@/components/ui/button';
 import {
   Dialog,
